Parser 12점 구현

+++ 컴퓨터공학창의설계/parser/replace_tmp.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 
-int g_value = 10;
+double g_value = 10;
 
 void func1() {
 	printf("I'm function 1\n");
@@ -11,14 +11,14 @@ void func2() {
 	printf("I'm function 2\n");
 }
 
-int g_dbl_value = 20;
+double g_dbl_value = 20;
 
 void main() {
-	int data1 = 10;
-	int data2 = 20;
-	int data3 = data1 * data2;
+	double data1 = 10;
+	double data2 = 20;
+	double data3 = data1 * data2;
 
-	int dbl_data = 1.1;
+	double dbl_data = 1.1;
 
 
 	func1();
